discord/requirements.txt
backports.tarfile==1.2.0
chill==0.10.0
discord==2.3.2
importlib-metadata==8.0.0
jaraco.collections==5.1.0
packaging==24.2
pip-chill==1.0.3
platformdirs==4.2.2
python-dotenv==1.0.1
tomli==2.0.1
